[test] remove extra setUpEnv/tearDownEnv
The setUpEnv and tearDownEnv methods are configured to run on
setUpClass/tearDownClass. Some testing systems as trial don't support that,
and that seems to cause some confusion in some systems, causing double
attempts to setup or teardown, and thus failing intermitently.

This commit removes the extra call to setUpEnv and tearDownEnv on tests on
this repository, as it now uses py.test to run tests, which supports
setUpClass and tearDownClass.