[bug] don't remove combined_bundle file on destructor
After PR #116, it's possible to inject an existing combined_ca_bundle
to the keymanager initialisation. However, it won't work for multiple
users because the file is removed when the keymanager is destroyed.

Considering there's no sensitive data on the file and it's created
as a temp file, we don't need to handle its deletion.